// packages/beam/src/types.ts
export type Scalar = 'f32' | 'i32' | 'u32'
export type VecType = 'vec2' | 'vec3' | 'vec4'
export type MatType = 'mat2' | 'mat3' | 'mat4'
export type NumType = Scalar | VecType | MatType
// ...
export type VertexSchema = Record<string, VecType | Scalar>
export type UniformSchema = Record<string, NumType>